refactor: Split view controller code.
