+++ kDrive/UI/Controller/Files/Save File/SaveFileViewController.swift
@@ -181,7 +181,7 @@ class SaveFileViewController: UIViewController {
         }
     }
 
-    private func dismiss(animated: Bool, clean: Bool = true, completion: (() -> Void)? = nil) {
+    func dismiss(animated: Bool, clean: Bool = true, completion: (() -> Void)? = nil) {
         Task {
             // Cleanup file that were duplicated to appGroup on extension mode
             if Bundle.main.isExtension && clean {
@@ -486,78 +486,4 @@ extension SaveFileViewController: SelectPhotoFormatDelegate {
             }
         }
     }
-}
-
-// MARK: - FooterButtonDelegate
-
-extension SaveFileViewController: FooterButtonDelegate {
-    @objc func didClickOnButton(_ sender: AnyObject) {
-        guard let drive = selectedDriveFileManager?.drive,
-              let directory = selectedDirectory else {
-            return
-        }
-
-        // Making sure the user cannot spam the button on tasks that may take a while
-        let button = sender as? IKLargeButton
-        button?.setLoading(true)
-
-        let items = items
-        guard !items.isEmpty else {
-            dismiss(animated: true)
-            return
-        }
-
-        Task {
-            await presentSnackBarSaveAndDismiss(files: items, directory: directory, drive: drive)
-        }
-    }
-
-    private func presentSnackBarSaveAndDismiss(files: [ImportedFile], directory: File, drive: Drive) async {
-        let message: String
-        do {
-            try await processForUpload(files: files, directory: directory, drive: drive)
-
-            message = files.count > 1 ? KDriveResourcesStrings.Localizable
-                .allUploadInProgressPlural(files.count) : KDriveResourcesStrings.Localizable
-                .allUploadInProgress(files[0].name)
-        } catch {
-            message = error.localizedDescription
-        }
-
-        Task { @MainActor in
-            self.dismiss(animated: true, clean: false) {
-                UIConstants.showSnackBar(message: message)
-            }
-        }
-    }
-
-    private func processForUpload(files: [ImportedFile], directory: File, drive: Drive) async throws {
-        try await fileImportHelper.saveForUpload(files, in: directory, drive: drive)
-        #if ISEXTENSION
-        showOpenAppToContinueNotification()
-        #endif
-    }
-
-    #if ISEXTENSION
-    //  Dynamic hook to be able to open an URL within an extension
-    @objc func openURL(_ url: URL) -> Bool {
-        var responder: UIResponder? = self
-        while responder != nil {
-            if let application = responder as? UIApplication {
-                return application.perform(#selector(openURL(_:)), with: url) != nil
-            }
-            responder = responder?.next
-        }
-        return false
-    }
-
-    func showOpenAppToContinueNotification() {
-        guard openURL(URLConstants.kDriveRedirection.url) else {
-            // Fallback on a local notification if failure to open URL
-            @InjectService var notificationHelper: NotificationsHelpable
-            notificationHelper.sendPausedUploadQueueNotification()
-            return
-        }
-    }
-    #endif
 }
